# components/sessionsrv/api/accounts/accounts.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "components/sessionsrv/api/accounts/accounts.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/accounts": {
      "post": {
        "operationId": "CreateAccount",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/responseAccountResp"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requestCreateAccountReq"
            }
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    },
    "/accounts/{username}": {
      "get": {
        "operationId": "GetAccount",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/responseAccountResp"
            }
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Accounts"
        ]
      },
      "patch": {
        "operationId": "UpdateAccount",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/responseAccountResp"
            }
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requestUpdateAccountReq"
            }
          }
        ],
        "tags": [
          "Accounts"
        ]
      }
    }
  },
  "definitions": {
    "accountsresponseAccount": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "title": "TODO - this shouldn't be duplicated with the request"
    },
    "requestCreateAccountReq": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "requestUpdateAccountReq": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "responseAccountResp": {
      "type": "object",
      "properties": {
        "account": {
          "$ref": "#/definitions/accountsresponseAccount"
        }
      }
    }
  }
}